Use new biasRNG function instead of stubbing.
The reason is stubbing replaced the whole method with a new, stubbed
method. However, that's not what I want. I only want to match certain
arguments; if there are no matches, then the function should fall
through to the real function.

There may be some tests I missed that depend on subtle chances.